build: move subprocess code under modules
Move subprocess.[ch] out of libutil so it can link against libzio in
the future. Following convention we put the new library under

 src/modules/libsubprocess

and also move unit test accordingly.
