+++ src/broker/Makefile.am
@@ -31,6 +31,7 @@ flux_broker_SOURCES = \
 	shutdown.c
 
 flux_broker_LDADD = \
+	$(top_builddir)/src/modules/libsubprocess/libsubprocess.la \
 	$(top_builddir)/src/modules/kvs/libkvs.la \
 	$(top_builddir)/src/common/libflux-core.la \
 	$(top_builddir)/src/common/libflux-internal.la
